Cameron Program v3
This is the work I have done as of 20/05/17
Minor tweak to the guess_add function, I have gotten rid of the .insert
method of adding the users guess to the blanks list. This is because by
inserting a letter, it adds an extra element which needs to be deleted
which in doing messes up the index of the letters.